+++ server/services/itemService.js
@@ -1,3 +1,4 @@
+import { NotFoundError } from "../middleware/errors/errorClasses.js";
 import Item from "../models/itemModel.js";
 import CategoryService from "./CategoryService.js";
 
@@ -7,15 +8,14 @@ class ItemService {
   }
 
   async getItemById(itemId) {
-    return await Item.findById(itemId);
+    const item = await Item.findById(itemId);
+    if (!item) throw new NotFoundError();
+    return item;
   }
 
   async addItem(data) {
     const newItem = new Item(data);
-    const category = await CategoryService.addItem(
-      data.category,
-      newItem._id
-    );
+    await CategoryService.addItem(data.category, newItem._id);
     return await newItem.save();
   }
 
@@ -23,13 +23,15 @@ class ItemService {
     const updatedItem = await Item.findByIdAndUpdate(itemId, data, {
       new: true,
     });
+    if (!updatedItem) throw new NotFoundError();
     return updatedItem;
   }
 
   async deleteItem(itemId) {
     const deletedItem = await Item.findByIdAndDelete(itemId);
-    const category = await CategoryService.removeItem(
-      deletedItem?.category,
+    if (!deletedItem) throw new NotFoundError();
+    await CategoryService.removeItem(
+      deletedItem.category,
       itemId
     );
     return deletedItem;
